+++ OpenEdXMobile/src/main/java/org/edx/mobile/player/IPlayerEventCallback.java
@@ -1,7 +1,5 @@
 package org.edx.mobile.player;
 
-import android.view.View;
-
 import java.io.Serializable;
 
 public interface IPlayerEventCallback extends Serializable {
@@ -9,5 +7,5 @@ public interface IPlayerEventCallback extends Serializable {
     void onError();
     void onPlaybackStarted();
     void onPlaybackComplete();
-    void saveCurrentPlaybackPosition(int currentPosition);
+    void saveCurrentPlaybackPosition(long currentPosition);
 }
