+++ OpenEdXMobile/src/main/java/org/edx/mobile/player/PlayerController.java
@@ -378,12 +378,12 @@ public class PlayerController extends FrameLayout {
         }
     }
 
-    private String stringForTime(int timeMs) {
-        int totalSeconds = timeMs / 1000;
+    private String stringForTime(long timeMs) {
+        long totalSeconds = timeMs / 1000;
 
-        int seconds = totalSeconds % 60;
-        int minutes = (totalSeconds / 60) % 60;
-        int hours   = totalSeconds / 3600;
+        long seconds = totalSeconds % 60;
+        long minutes = (totalSeconds / 60) % 60;
+        long hours   = totalSeconds / 3600;
 
         mFormatBuilder.setLength(0);
         if (hours > 0) {
@@ -393,13 +393,13 @@ public class PlayerController extends FrameLayout {
         }
     }
 
-    private synchronized int setProgress() {
+    private synchronized long setProgress() {
         if (mPlayer == null || mDragging) {
             return 0;
         }
 
-        int position = mPlayer.getCurrentPosition();
-        int duration = mPlayer.getDuration();
+        long position = mPlayer.getCurrentPosition();
+        long duration = mPlayer.getDuration();
         if (mProgress != null) {
             if (duration > 0) {
                 // use long to avoid overflow
@@ -713,7 +713,7 @@ public class PlayerController extends FrameLayout {
             if (mPlayer == null) {
                 return;
             }
-            int pos = mPlayer.getCurrentPosition();
+            long pos = mPlayer.getCurrentPosition();
             try{
                 mPlayer.callPlayerSeeked(pos, pos-30000, true);
             }catch(Exception e){
@@ -791,7 +791,7 @@ public class PlayerController extends FrameLayout {
                     return;
                 }
 
-                int pos;
+                long pos;
                 switch (msg.what) {
                 case FADE_OUT:
                     if (view.mIsAutoHide) {
